refactor stock summary code to make it more readable
