+++ apps/client/app/helpers/stock-summary.js
@@ -1,0 +1,16 @@
+export const fetchStockItems = async () => {
+    const response = await fetch('http://localhost:3100/api/stock-summaries');
+    const data = await response.json();
+    return data;
+  };
+  export function accumulateStockItems(stockItems) {
+    return stockItems.reduce((acc, item) => {
+        if (acc.has(item.itemId)) {
+          const existingItem = acc.get(item.itemId);
+          acc.set(item.itemId, { ...item, SOH: existingItem.SOH + item.SOH });
+        } else {
+          acc.set(item.itemId, item);
+        }
+        return acc;
+      }, new Map());
+  }
