+++ apps/client/app/stock-summary/page.tsx
@@ -2,6 +2,8 @@
 import { useEffect, useState } from 'react';
 import { Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Paper } from '@mui/material';
 import SOHByBranch from '../components/Stock/StockOnHandPopUp';
+import { fetchStockItems } from '../helpers/stock-summary';
+import { accumulateStockItems } from '../helpers/stock-summary';
 
 interface StockItem {
   id: number;
@@ -21,28 +23,18 @@ export default  function Index()  {
   const [selectedStock, setSelectedStock] = useState<StockItem | null>(null);
 
   useEffect(() => {
-    // Simulate fetching data from an API
-    const fetchStockItems = async () => {
-      const response = await fetch('http://localhost:3100/api/stock-summaries'); 
-      const data = await response.json();
+    const fetchStock = async () => {
+      const data = await fetchStockItems();
       setStockItems(data);
     };
 
-    fetchStockItems();
+    fetchStock();
   }, []);
 
   const handleSOHClick = (stock: StockItem) => {
     setSelectedStock(stock);
   };
-  const accumulatedStockItemsMap = stockItems.reduce((acc, item) => {
-    if (acc.has(item.itemId)) {
-      const existingItem = acc.get(item.itemId);
-      acc.set(item.itemId, { ...item, SOH: existingItem.SOH + item.SOH });
-    } else {
-      acc.set(item.itemId, item);
-    }
-    return acc;
-  }, new Map());
+  const accumulatedStockItemsMap = accumulateStockItems(stockItems);
 
   const uniqueStockItems = Array.from(accumulatedStockItemsMap.values());
